break PBJVideoPlayerController into two classes, PBJVideoPlayer which handles all the actual player interaction, and PBJVideoPlayerController which is a UIViewController that offers a slight wrapper around PBJVideoPlayer to support things like touch interactions.
This is a breaking change, but still provides all of the original functionality, just in slightly different places. Porting old code to the new interface should be straightforward, as almost everything that's changed is simply going to now reference the 'player' property of the PBJVideoPlayerController for the functionality that used to be on PBJVideoPlayerController itself.

This change allows the player to be used without being a view controller.
