+++ Source/PBJVideoPlayer.h
@@ -1,0 +1,74 @@
+//
+//  PBJVideoPlayerController.h
+//
+//  Created by Patrick Piemonte on 5/27/13.
+//  Copyright (c) 2013-present, Patrick Piemonte, http://patrickpiemonte.com
+//
+//  Permission is hereby granted, free of charge, to any person obtaining a copy of
+//  this software and associated documentation files (the "Software"), to deal in
+//  the Software without restriction, including without limitation the rights to
+//  use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+//  the Software, and to permit persons to whom the Software is furnished to do so,
+//  subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in all
+//  copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+//  FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+//  COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+//  IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+//  CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+//
+
+#import <UIKit/UIKit.h>
+#import "PBJVideoView.h"
+
+typedef NS_ENUM(NSInteger, PBJVideoPlayerPlaybackState) {
+    PBJVideoPlayerPlaybackStateStopped = 0,
+    PBJVideoPlayerPlaybackStatePlaying,
+    PBJVideoPlayerPlaybackStatePaused,
+    PBJVideoPlayerPlaybackStateFailed,
+};
+
+typedef NS_ENUM(NSInteger, PBJVideoPlayerBufferingState) {
+    PBJVideoPlayerBufferingStateUnknown = 0,
+    PBJVideoPlayerBufferingStateReady,
+    PBJVideoPlayerBufferingStateDelayed,
+};
+
+// PBJVideoPlayerController.view provides the interface for playing/streaming videos
+@protocol PBJVideoPlayerDelegate;
+
+@interface PBJVideoPlayer : NSObject
+
+@property (nonatomic, weak) id<PBJVideoPlayerDelegate> delegate;
+
+@property (nonatomic, copy) NSString *videoPath;
+@property (nonatomic, copy) NSString *videoFillMode; // default, AVLayerVideoGravityResizeAspect
+@property (nonatomic, readonly) PBJVideoView * videoView;
+
+@property (nonatomic) BOOL playbackLoops;
+@property (nonatomic) BOOL playbackFreezesAtEnd;
+@property (nonatomic, readonly) PBJVideoPlayerPlaybackState playbackState;
+@property (nonatomic, readonly) PBJVideoPlayerBufferingState bufferingState;
+
+@property (nonatomic, readonly) NSTimeInterval maxDuration;
+
+- (void)playFromBeginning;
+- (void)playFromCurrentTime;
+- (void)pause;
+- (void)stop;
+
+@end
+
+@protocol PBJVideoPlayerDelegate <NSObject>
+@required
+- (void)videoPlayerReady:(PBJVideoPlayer *)videoPlayer;
+- (void)videoPlayerPlaybackStateDidChange:(PBJVideoPlayer *)videoPlayer;
+
+- (void)videoPlayerPlaybackWillStartFromBeginning:(PBJVideoPlayer *)videoPlayer;
+- (void)videoPlayerPlaybackDidEnd:(PBJVideoPlayer *)videoPlayer;
+
+@end
